+++ oauth/__init__.py
@@ -1,0 +1,2 @@
+#from .models import AccessToken
+default_app_config = "oauth.apps.OauthConfig"